+++ src/netgate_xml_to_xlsx/formats/__init__.py
@@ -1,7 +1,6 @@
 """Formats module."""
 # Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
 
-from .html import HtmlFormat  # NOQA
 from .text import TextFormat  # NOQA
 from .xlsx import XlsxFormat  # NOQA